+++ src/main/kotlin/com/atlassian/performance/tools/aws/api/BatchingCloudformation.kt
@@ -5,6 +5,7 @@ import net.jcip.annotations.ThreadSafe
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
 import java.time.Duration
+import java.util.*
 import java.util.concurrent.CompletableFuture
 import java.util.concurrent.ConcurrentHashMap
 import kotlin.concurrent.timer
@@ -13,16 +14,18 @@ import kotlin.concurrent.timer
  * Batches requests to AWS CloudFormation to avoid request throttling.
  *
  * Reuse instances of [BatchingCloudformation] and maximize concurrency to get the most benefits of batching.
+ *
+ * @param [refreshPeriod] Gives time for requests to accumulate before sending them as a batch.
  */
 @ThreadSafe
-class BatchingCloudformation(
-    private val cloudformation: ScrollingCloudformation
+class BatchingCloudformation internal constructor(
+    private val cloudformation: ScrollingCloudformation,
+    refreshPeriod: Duration
 ) {
     private val logger: Logger = LogManager.getLogger(this::class.java)
     private val namesToStackRequests: MutableMap<String, CompletableFuture<Stack?>> = ConcurrentHashMap()
-    private val refreshPeriod = Duration.ofMinutes(1)
 
-    init {
+    private val timer: Lazy<Timer> = lazy {
         timer(
             name = "batching-cloudformation-refresh",
             daemon = true,
@@ -31,11 +34,24 @@ class BatchingCloudformation(
         )
     }
 
+    @Deprecated(
+        message = "The constructor wasn't supposed to become API. Use the instance exposed by Aws instead.",
+        replaceWith = ReplaceWith("aws.batchingCloudformation")
+    )
+    constructor(
+        cloudformation: ScrollingCloudformation
+    ) : this(
+        cloudformation = cloudformation,
+        refreshPeriod = Duration.ofMinutes(1)
+    )
+
     fun findStack(
         stackName: String
     ): CompletableFuture<Stack?> {
         synchronized(namesToStackRequests) {
-            return namesToStackRequests.getOrPut(stackName) { CompletableFuture() }
+            val stackRequest = namesToStackRequests.getOrPut(stackName) { CompletableFuture() }
+            timer.value
+            return stackRequest
         }
     }